Detect I2C addresses
That's one small step for the code, one massive change for operational efficiency! Up till now, it has been necessary to recompile every time I've changed the I2C device configuration. But now, the code checks what's connected each time it runs, and runs the appropriate code. That's especially helpful with the BMP280 sensors being on 0x76 or 0x77 depending on which board I'm plugging in...